refactor : refactor for menugroup packaging
메뉴그룹 관련 패키징 변경
